Wait for the out-of-process notification worker to launch before running tests
Launching the worker is a substantial part of the test's total runtime, and
sometimes took long enough that it produced spurious timeouts. This also tweaks
the worker launching logic to have it avoid launching more workers when there's
enough to serve the entire queue, but some simply haven't finished starting.